# openapi.yaml
components:
...
  responses:
    BadRequest:
      content:
        application/json:
          example:
            code: 400
            message: Malformed parameters could not be processed
          schema:
            $ref: "#/components/schemas/response"
      description: Bad Request
    Forbidden:
      content:
        application/json:
          example:
            code: 403
            message: You are not authorized to access this ressource
          schema:
            $ref: "#/components/schemas/response"
      description: Forbidden
...
    NotFound:
      content:
        application/json:
          example:
            code: 404
            message: User not found
          schema:
            $ref: "#/components/schemas/response"
      description: Not Found
    Success:
      content:
        application/json:
          example:
            code: 200
            message: Successful operation
          schema:
            $ref: "#/components/schemas/response"
      description: Signals the operation was successful
    Unauthorized:
      content:
        application/json:
          example:
            code: 401
            message: You are not authorized to access this ressource.
          schema:
            $ref: "#/components/schemas/response"
      description: Unauthorized (should redirect on login screen)
  schemas:
...
    credentials:
      description: the registering user informations to validate.
      properties:
        password:
          $ref: "#/components/schemas/password"
        username:
          $ref: "#/components/schemas/username"
      type: object
...
    password:
      description: |
        the user's password. It must be at least 12 characters long, and contain:
        - at least one **uppercase letter** [A-Z]
        - at least one **lowercase letter** [a-z]
        - at least one **digit** [0 - 9]
        - at least one **special characters** [@ $ ! % * ? &]
      example: $FillinglessPie1249%
      format: password
      maxLength: 256
      minLength: 12
      pattern: ^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&]).{12,}$
      type: string
...
    response:
      description: an object describing an HTTP response, to be referenced in responses schemas.
      properties:
        code:
          description: the value of the response code
          example: 418
          format: integer
          type: number
        details:
          description: more information on the response
          example: I'm a teapot !
          format: plaintext
          maxLength: 256
          type: string
        message:
          description: a human-readable message describing the response
          example: I cannot make coffee.
          format: plaintext
          maxLength: 256
          type: string
      required:
        - code
        - message
      type: object
...
    username:
      description: |
        the user's chosen username, used as an identifier within the system. <br>
        It is an alphanumeric string comprised between 4 and 18 characters.
      example: "Dandelion"
      format: plaintext
      maxLength: 18
      minLength: 4
      pattern: "^[a-zA-Z0-9]{4,18}$"
      type: string
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: the authentication informations used by CLI users. Conforms with RFC8725.
      scheme: bearer
      type: http
    cookieAuth:
      description: the authentication informations used by browser users.
      in: cookie
      name: jwt
      type: apiKey
info:
  contact:
    email: cdomet-d@student.42lyon.French
    name: cdomet-d
    url: https://localhost:8443/support
  description: Documentation for the RESTful section the BigTeam's Transcendance project
  license:
    name: MIT
    url: https://mit-license.org/
  title: BigTeam API
  version: 1.0.0
openapi: 3.1.0
paths:
  /api/account:
    delete:
      description: |
        Deletes an user account and all associated information. <br>
        The user will be prompted to send their credentials to validate their identity before deletion can happen.
      operationId: deleteAccount
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - cookieAuth: []
        - bearerAuth: []
      summary: deletes an authenticated user account
      tags:
        - account management
    post:
      description: |
        Creates a new persistent or temporary user
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/credentials"
        description: the user's username and password, used to create the account.
        required: true
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
      summary: registers a new user
      tags:
        - account management
...
servers:
  - description: Our lone server, we don't really need a production one
    url: https://localhost:8443
...